demo/requirements.txt
einops==0.6.0
nltk==3.5
numpy==1.24.3
python-levenshtein==0.20.9
PyYAML==6.0
timm==0.6.13
--extra-index-url https://download.pytorch.org/whl/cu113
torch==2.0.0
# torch==2.0.0
--extra-index-url https://download.pytorch.org/whl/cu113
torchvision==0.15.1
# torchvision==0.15.1
pdf2image
scipy==1.10.1
open-fonts==0.1
natsort==8.2.0
levenshtein==0.20.9
distance==0.1.3
opencv-python-headless==4.7.0.72
pillow==9.4.0
scikit-image==0.19.3
albumentations==1.3.0
streamlit==1.22.0
